fix(vps): confirmInvite as a real HttpApiEndpoint, not a raw route
Adversarial review on this PR caught a real architectural mistake: the
claim that HttpApiBuilder.group handlers can't set response headers was
wrong. effect@4.0.0-beta.93's HttpApiBuilder.ts (handlerToHttpEffect)
checks Response.isHttpServerResponse(response) on the handler's return
value and short-circuits schema encoding if true, passing a
handler-constructed HttpServerResponse straight through -- headers
included. This means confirmInvite never needed the raw HttpRouter.add
exception; the handler can build the response with HttpServerResponse.json
+ setHeader directly and return it.

Converts confirmInvite to a real HttpApiEndpoint (packages/api/src/invite.ts),
merges its handler into invite.handlers.ts's HttpApiBuilder.group chain,
and deletes invite-confirm.route.ts entirely. Both invite endpoints now
go through the same typed contract instead of one being a special case --
removes a false precedent that would have misled later steps with similar
cookie-setting needs into believing the raw-router escape hatch was
necessary.
